sRGBProfile implemented

+++ ColorKit/Classes/AbsoluteColorSpaces/XYZSpace.swift
@@ -5,7 +5,6 @@
 //  Created by Lorenzo Bernaschina on 13/02/2019.
 //
 
-import Foundation
 
 internal protocol AbsoluteSpace {
     func fromXYZSpace(space: XYZSpace)
@@ -13,10 +12,9 @@ internal protocol AbsoluteSpace {
 }
 
 internal class XYZSpace {
-    
-    public var X: CGFloat
-    public var Y: CGFloat
-    public var Z: CGFloat
+    public var X: Float
+    public var Y: Float
+    public var Z: Float
     
     public init() {
         self.X = 0.0
@@ -24,7 +22,7 @@ internal class XYZSpace {
         self.Z = 0.0
     }
     
-    public init(X: CGFloat, Y: CGFloat, Z: CGFloat) {
+    public init(X: Float, Y: Float, Z: Float) {
         self.X = X
         self.Y = Y
         self.Z = Z
